ai-compose/ai-model/requirements.txt
black
fastapi
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
python-dotenv
ruff
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
uvicorn